Fix officers view
